Move basebackup files into a separate folder
